:sparkles: Feat: DiaryResponseConverter 구현
<body>
- 기록 응답 Dto builder 컨버터 구현

- 관련 : #270

+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/record/dto/DiaryResponse.java
@@ -14,7 +14,7 @@ public class DiaryResponse {
 	public static class DiaryDto {
 		private Long diaryId;
 		private String content;
-		private Integer enjoyRating;
+		private double enjoyRating;
 		private List<DiaryImageDto> diaryImages;
 	}
